major refactor - moved utils to module, backup for footprints file type for gpkg
