+++ 03_MoveProducts.py
@@ -8,7 +8,7 @@ import warnings
 
 import rasterio
 
-from processing_utils import *
+from src.macs_processing.utils.processing import *
 from utils_postprocessing import *
 
 warnings.filterwarnings("ignore")
