+++ src/macs_processing/__init__.py
@@ -1,0 +1,2 @@
+def hello() -> str:
+    return "Hello from macs-processing-2023!"
